2_model_training_sproc/requirements.txt
snowflake-snowpark-python
snowflake-ml-python
toml # for local development
pandas
cachetools
numpy


